split insert and update
